+++ src/remix/sdk-setup.ts
@@ -73,6 +73,7 @@ interface SdkAstOptions {
   replaysSessionSampleRate?: number;
   replaysOnErrorSampleRate?: number;
   integrations?: Array<Proxified>;
+  enableLogs?: boolean;
 }
 
 function getInitCallArgs(
@@ -81,6 +82,7 @@ function getInitCallArgs(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
 ) {
   const initCallArgs: SdkAstOptions = {
@@ -90,6 +92,11 @@ function getInitCallArgs(
   // Adding tracing sample rate for both client and server
   if (selectedFeatures.performance) {
     initCallArgs.tracesSampleRate = 1.0;
+  }
+
+  // Adding logs for both client and server
+  if (selectedFeatures.logs) {
+    initCallArgs.enableLogs = true;
   }
 
   // Adding integrations and replay options only for client
@@ -132,6 +139,7 @@ function insertClientInitCall(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
 ): void {
   const initCallArgs = getInitCallArgs(dsn, 'client', selectedFeatures);
@@ -155,6 +163,7 @@ export function generateServerInstrumentationFile(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
 ) {
   // create an empty file named `instrument.server.mjs`
@@ -192,6 +201,7 @@ export async function createServerInstrumentationFile(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
 ) {
   const { instrumentationFile, instrumentationFileMod } =
@@ -207,6 +217,7 @@ export async function insertServerInstrumentationFile(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
 ) {
   const instrumentationFile = await createServerInstrumentationFile(
@@ -361,6 +372,7 @@ export function updateEntryClientMod(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ): ProxifiedModule<any> {
@@ -407,6 +419,7 @@ export async function initializeSentryOnEntryClient(
   selectedFeatures: {
     performance: boolean;
     replay: boolean;
+    logs: boolean;
   },
 ): Promise<void> {
   const clientEntryFilename = `entry.client.${isTS ? 'tsx' : 'jsx'}`;